requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
gql==2.0.0
graphql-core==2.3.2
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
infura==0.2.1
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.5
Pillow==8.2.0
plotly==4.14.3
promise==2.3
psycopg2==2.8.6
pycodestyle==2.6.0
Pygments==2.8.1
pylint==2.6.0
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-cache==0.5.2
requests-toolbelt==0.9.1
retrying==1.3.3
Rx==1.6.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
websocket-client==0.57.0
whitenoise==5.2.0
wrapt==1.12.1
